Get all photos instead of using adapter.refresh()
- As described here: https://stackoverflow.com/q/63985957/4612653
" From the details screen, pressing the back button returns the user to
the gallery. If the users pulls the layout, the progress indicator
appears but adapter.refresh() does not happen."
- Instead of using adapter.refresh(), I would just trigger a fresh call
to get all photos. It seems that for adapter.refresh() to work, a new
PagingSource has to be created, which only happens when a call to
getAllPhotos is made.

Signed-off-by: Alvin Dizon <alvin.dizon91@gmail.com>